// src/types/NewInstanceConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RuntimeSource } from "./RuntimeSource";

export type NewInstanceConfig = { runtime: RuntimeSource, instanceName: string, instanceDescription: string, instanceRootPath: string, instanceDataDir: string, serverHostName: string, maxPlayers: number, serverPort: number, password: string, masterServerEnabled: boolean, };